refactor: split syslog service module
