+++ src/mcp.rs
@@ -1,4 +1,4 @@
-use crate::app::LogService;
+use crate::app::SyslogService;
 use crate::config::McpConfig;
 
 mod protocol;
@@ -11,6 +11,6 @@ pub use routes::router;
 /// Shared app state
 #[derive(Clone)]
 pub struct AppState {
-    pub service: LogService,
+    pub service: SyslogService,
     pub config: McpConfig,
 }
